refactor(chat-provider): reorganize provider and base classes into dedicated modules
This commit restructures the codebase to improve modularity and maintainability by:

- Moving the `Reasoner` base class and provider implementations from `tools.internal` to dedicated modules
- Creating new top-level modules: `providers/` for LLM provider implementations and `tools/base.py` for the base `Reasoner` class
- Removing deprecated `tools.internal` module and its contents
- Updating all imports to use the new module structure
- Preserving all functionality while improving code organization and extensibility

The change enables better separation of concerns, making it easier to add new providers (e.g., OpenAI, Anthropic) and maintain the core reasoning logic independently.
