+++ src/hcaptcha_challenger/providers/__init__.py
@@ -1,5 +1,9 @@
 # -*- coding: utf-8 -*-
-# Provider implementations for different LLM backends.
+"""
+Providers - LLM provider implementations.
+
+This module provides provider implementations for different LLM backends.
+"""
 
 from .protocol import ChatProvider
 from .gemini import GeminiProvider
