--- src/hcaptcha_challenger/tools/internal/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-# Internal implementation module for tools package.
-# External code should NOT import from this module directly.
-
-from .base import Reasoner
-
-__all__ = ["Reasoner"]
